requirements-adk.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
numpy==1.26.0
pandas==2.1.1
python-docx==1.0.1
PyMuPDF==1.23.3
python-pptx==0.6.22
Pillow==9.5.0
chromadb==0.4.18
sentence-transformers==2.2.2
google-generativeai==0.3.1
tqdm==4.66.1
scikit-learn==1.3.2
pyngrok==7.0.0
requests==2.31.0
psutil==5.9.6
streamlit==1.28.0
matplotlib==3.8.0
python-dotenv==1.0.0
# ADK and dependencies
google-adk==0.2.0
litellm>=0.12.0
